Initial work has been implemented
The structure of productCol and map.js has been reworked.
The sellersColumn is no longer indexed, there is now instead a hashMap that holds a key and a value.

the key is the productID, the key gives the information about the product.
By doing this the marker now knows what defines it (information about the product)